+++ example/fill.py
@@ -12,25 +12,21 @@ import frykit.shp as fshp
 provinces = fshp.get_cn_province()
 data = np.linspace(0, 100, len(provinces))
 
-# 设置地图范围
-extents1 = [78, 134, 14, 55]
-extents2 = [105, 120, 2, 25]
-
 # 设置投影
 map_crs = fplt.CN_AZIMUTHAL_EQUIDISTANT
 data_crs = ccrs.PlateCarree()
 
 # 准备主地图
 fig = plt.figure(figsize=(10, 6))
-ax1 = fig.add_subplot(projection=map_crs)
-ax1.set_extent(extents1, crs=data_crs)
-fplt.add_nine_line(ax1, lw=0.5)
-ax1.axis('off')
+main_ax = fig.add_subplot(projection=map_crs)
+main_ax.set_extent((78, 134, 14, 55), data_crs)
+fplt.add_nine_line(main_ax, lw=0.5)
+main_ax.axis('off')
 
 # 准备小地图
-ax2 = fplt.add_mini_axes(ax1)
-ax2.set_extent(extents2, crs=data_crs)
-fplt.add_nine_line(ax2, lw=0.5)
+mini_ax = fplt.add_mini_axes(main_ax)
+mini_ax.set_extent((105, 120, 2, 25), data_crs)
+fplt.add_nine_line(mini_ax, lw=0.5)
 
 # 填色参数
 bins = [0, 20, 40, 60, 80, 100]
@@ -38,7 +34,7 @@ colors = ['#dbdee7', '#afc6e8', '#7a9cdc', '#3b6cb8', '#2a3b97']
 nbin = len(bins) - 1
 labels = [f'{bins[i]} - {bins[i + 1]}' for i in range(nbin)]
 
-# 准备 colormap 和 norm
+# 准备 cmap 和 norm
 norm = mcolors.BoundaryNorm(bins, nbin)
 cmap = mcolors.ListedColormap(colors)
 
@@ -46,7 +42,7 @@ cmap = mcolors.ListedColormap(colors)
 path_effects = [Stroke(linewidth=1.5, foreground='w'), Normal()]
 
 # 绘制填色多边形，标注省名。
-for ax in [ax1, ax2]:
+for ax in [main_ax, mini_ax]:
     fplt.add_geoms(
         ax, provinces, array=data, cmap=cmap, norm=norm, ec='k', lw=0.4
     )
@@ -60,7 +56,8 @@ patches = []
 for color, label in zip(colors, labels):
     patch = mpatches.Patch(fc=color, ec='k', lw=0.5, label=label)
     patches.append(patch)
-ax1.legend(
+
+main_ax.legend(
     handles=patches,
     loc=(0.05, 0.05),
     frameon=False,
@@ -70,8 +67,8 @@ ax1.legend(
 )
 
 # 添加指北针和比例尺
-fplt.add_compass(ax1, 0.5, 0.85, size=20, style='circle')
-scale_bar = fplt.add_scale_bar(ax1, 0.22, 0.1, length=1000)
+fplt.add_compass(main_ax, 0.5, 0.85, size=20, style='circle')
+scale_bar = fplt.add_scale_bar(main_ax, 0.22, 0.1, length=1000)
 scale_bar.set_xticks([0, 500, 1000])
 scale_bar.set_xticks([250, 750], minor=True)